requirements.txt
flask_cors
joblib
requests
bs4
googlesearch-python
scikit-learn